+++ src/Tile/Tile.test.js
@@ -1,4 +1,4 @@
-import { Button } from '../Button/Button';
+import { Button } from '../';
 import { Identifier } from '../Identifier/Identifier';
 import { mount } from 'enzyme';
 import { Popover } from '../Popover/Popover';
